Point pre-prod and dev environments to distribution-dev environment
